chore: move functions into utils
